chore: refactor project structure
- Project structure is moved to feature based domain-driven design to be more scalable and maintainable for the future
