+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "vue": "^3.5.13",
+    "vue-router": "^4.5.0",
     "vuex": "^4.1.0"
   },
   "devDependencies": {
