+++ src/features/horses/components/HorseTableHeader.vue
@@ -1,0 +1,15 @@
+<template>
+  <thead class="bg-gray-100 border-b sticky top-0 z-10">
+    <tr>
+      <th class="px-4 py-2 font-semibold">Name</th>
+      <th class="px-4 py-2 font-semibold">Condition</th>
+      <th class="px-4 py-2 font-semibold">Color</th>
+    </tr>
+  </thead>
+</template>
+
+<script setup lang="ts">
+</script>
+
+<style scoped>
+</style>
